backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
passlib[bcrypt]
python-jose
python-multipart
python-dotenv
requests
openai